+++ demo/config/environments/test.rb
@@ -43,4 +43,5 @@ Rails.application.configure do
   # Raises error for missing translations.
   # config.action_view.raise_on_missing_translations = true
   config.primer_view_components.silence_deprecations = true
+  config.primer_view_components.force_system_arguments = false
 end