+++ demo/test/components/previews/primer/docs/tab_nav_component_preview.rb
@@ -8,6 +8,8 @@ module Primer
       def with_panels; end
 
       def with_extra_content; end
+
+      def adding_extra_content_after_the_tabs; end
     end
   end
 end