+++ test/lib/classify_test.rb
@@ -16,6 +16,8 @@ class PrimerClassifyTest < Minitest::Test
     assert_generated_class("f4",  { font_size: 4 })
     assert_generated_class("f5",  { font_size: 5 })
     assert_generated_class("f6",  { font_size: 6 })
+    assert_generated_class("text-small", { font_size: :small })
+    assert_generated_class("text-normal",  { font_size: :normal })
   end
 
   def test_m
@@ -359,10 +361,23 @@ class PrimerClassifyTest < Minitest::Test
     assert_generated_class("text-left",       { text_align: :left })
   end
 
+  def test_font_family
+    assert_generated_class("text-mono", { font_family: :mono })
+  end
+
+  def test_font_style
+    assert_generated_class("text-italic", { font_style: :italic })
+  end
+
+  def test_text_transform
+    assert_generated_class("text-uppercase", { text_transform: :uppercase })
+  end
+
   def test_font_weight
-    assert_generated_class("text-light",    { font_weight: :light })
-    assert_generated_class("text-normal",   { font_weight: :normal })
-    assert_generated_class("text-bold",     { font_weight: :bold })
+    assert_generated_class("text-light",      { font_weight: :light })
+    assert_generated_class("text-normal",     { font_weight: :normal })
+    assert_generated_class("text-bold",       { font_weight: :bold })
+    assert_generated_class("text-emphasized", { font_weight: :emphasized })
   end
 
   def test_box_shadow
@@ -615,25 +630,13 @@ class PrimerClassifyTest < Minitest::Test
     assert_generated_class("hover-grow", { animation: :grow })
   end
 
-  def test_raises_error_when_passing_in_a_primer_css_class_name_in_development
-    ENV["RAILS_ENV"] = "development"
-    exception = assert_raises ArgumentError do
-      Primer::Classify.call(classes: "bg-blue text-center float-left ml-1")
-    end
-
-    assert_includes exception.message, "Primer CSS class names"
-  ensure
-    ENV["RAILS_ENV"] = "test"
-  end
-
   def test_raises_does_not_raise_error_when_passing_in_a_primer_css_class_name_in_development_and_flag_is_set
-    ENV["RAILS_ENV"] = "development"
     ENV["PRIMER_WARNINGS_DISABLED"] = "1"
-
-    assert_generated_class("bg-blue text-center float-left ml-1", { classes: "bg-blue text-center float-left ml-1" })
+    with_force_system_arguments(true) do
+      assert_generated_class("bg-blue text-center float-left ml-1", { classes: "bg-blue text-center float-left ml-1" })
+    end
   ensure
     ENV["PRIMER_WARNINGS_DISABLED"] = nil
-    ENV["RAILS_ENV"] = "test"
   end
 
   def test_does_not_raise_error_when_passing_in_a_primer_css_class_otherwise
@@ -644,6 +647,22 @@ class PrimerClassifyTest < Minitest::Test
     generated_class = Primer::Classify.call(classes: "foo-class")[:class]
     refute(generated_class.start_with?(" "))
     refute(generated_class.end_with?(" "))
+  end
+
+  def test_raises_if_not_using_system_arguments_when_force_system_arguments_is_true
+    with_force_system_arguments(true) do
+      exception = assert_raises ArgumentError do
+        Primer::Classify.call(classes: "d-block")
+      end
+
+      assert_includes exception.message, "Use System Arguments (https://primer.style/view-components/system-arguments) instead of Primer CSS class"
+    end
+  end
+
+  def test_does_not_raise_if_not_using_system_arguments_when_force_system_arguments_is_false
+    with_force_system_arguments(false) do
+      assert_generated_class("d-block", { classes: "d-block" })
+    end
   end
 
   private