+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## main
 
+### New
+
+* Add `font_family`, `font_style` and `text_transform` system arguments.
+
+    *Manuel Puyol*
+
+* Add more options for `font_size` and `font_weight`.
+
+    *Manuel Puyol*
+
+### Updates
+
+* Add `align` option to the `TabNav` extra slot to allow HTML ordering.
+
+    *Manuel Puyol*
+
 ### Misc
 
 * Auto-generate component previews from doc examples and run integration test checks.
@@ -13,6 +29,10 @@
     *Kate Higa*
 
 * Only include `ViewComponent::SlotableV2` if `ViewComponent::Base` does not already include it.
+
+    *Manuel Puyol*
+
+* Add `force_system_arguments` option to raise an error if a class is used instead of using System Arguments.
 
     *Manuel Puyol*